Add experimental data
